Animations prep part II
Adds new hidden inputs so we can pass the pan and zoom state to the backend on render. With these it should be possible now to calculate the position of an element in the previous view on the backend so that we can then setup the animation for the new view.

Also does a couple of other things:

- just cache the mermaid build before setting attributes not after. This is important for the next step where we'll be mutating the SVG with the previous animation otherwise
- move the highlighted handling to the svg attribute setting. This imporves performance when clicking nodes massively