Cleans up paramerers and state storage in preparation for further animations work (#87)
* Cleans up paramerers and state storage in preparation for further animations work

The main change here is to cleanup parameters for different usages. Notably we now have

- RootParams - parameters for the root path. These are the URL parameters
- UpdateParams - includes RootParams plus the additional form parameters that can be passed and layout information
- GenerateParams - the params that affect the mermaid svg generation
- Session - parameters necessary to handle change logic between two states. Note we do not store layout information here as when we compare layouts we always want to use the latest layout.

This also tidies up url parameters so unused URL parameters like `shouldExpand` are no longer stored in the URL

The change includes one regression which is to remove expandedIds from the path. This is to ensure the expandedIds never generate a URL too long by becoming too numerous. They are now stored on the session. I'd suggest if we want these to be passable by a URL share we actually put the sessionId in the path as that's bounded but also potentially dangerous. I'd like to see if anyone notices the regression.

Lastly there's the first steps towards aniumations. We use an opacity tranform on the new svg to stop the movement of it popping. Then use a view-transformation-api transform to cross fade between the new and old. This works particularly well when only the highlighted node changes.

* Updating version to 0.0.78

* Fix parse/stringify bug

* Fit pan bug

* Increase test timeouts

* Remove console log

* comment

---------

Co-authored-by: dc-autobot[bot] <181364585+dc-autobot[bot]@users.noreply.github.com>